Started implementing a functional solution for button actions.
The idea is that the actions need to be stored in data, so there's going to be a large enum where each element is associated with some action.
But, it would be nice if we could compose these actions fairly nicely, so the enum returns a closure with composable types rather than performing the action.
However, currently getting a weird 'temporary value is freed' error, will look into it some more soon.