Moving away from prefabs and instead focusing on serializing levels.
I think I can worry about prefabs later once I start moving to 3d.
If I do it now, it'll just be a lot of tech debt.

As for serializing levels, I think I need to figure out a good way to Serialize the Progession.
Obviously it should be a vec of boxouts, but actually translating that to a progression could be tricky.