Created LevelData.
LevelData depends on ButtonData, and is used to make levels serializable.
It has many functions that operate on an amethyst World to insert a level.
You can try_from a level's name, and it'll try and deserialize it from a matching filename in assets/levels.
The current level was moved to this system.

Most of the code that was in black_state.rs is now in level.rs.
This code is still pretty messy and should be cleaned up eventually.
I also think we could probably modularize it further, using LevelData to only add the gameplay components, instead of everything.
This probably not a super high priority though.

The next thing I need to work on is making some sort of transition from level to level.
I'll need some way for the program to know the order of the levels as well.