Updated some levels in response to play tests.
Level 1 - I changed it so 4 and 5 are mapped properly.
It may just be the fact that currently the number keys are used to press buttons
or the fact that the output doesn't display properly on the first level because of window resizing
but the fact that the 4 and 5 are switched seems more like a confusing mistake than enlightening as it stands.
May want to change this back eventually.

Also changed level 7 and 8 to have more callbacks to previous level prompts.
This hopefully also accomplished two other things: making the level 7 prompt longer and the level 8 prompt more helpful.